+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing so far
+- Introduce options to control TLS and timeout configuration
+- Write unit tests
 
 ## [v0.3.0] - 2019-03-18
 - Update to the changed Module interface of joe v0.4.0
@@ -20,6 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial alpha release
 
 [Unreleased]: https://github.com/go-joe/http-server/compare/v0.3.0...HEAD
-[v0.2.0]: https://github.com/go-joe/http-server/compare/v0.2.0...v0.3.0
+[v0.3.0]: https://github.com/go-joe/http-server/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/go-joe/http-server/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/go-joe/http-server/releases/tag/v0.1.0